Develop

+++ textr/src/main/java/org/usth/ict/ulake/textr/services/IndexFilesService.java
@@ -2,6 +2,7 @@ package org.usth.ict.ulake.textr.services;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import io.smallrye.mutiny.Uni;
 import org.eclipse.microprofile.jwt.JsonWebToken;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 import org.slf4j.Logger;
@@ -12,7 +13,8 @@ import org.springframework.data.domain.Pageable;
 import org.usth.ict.ulake.common.model.LakeHttpResponse;
 import org.usth.ict.ulake.common.model.dashboard.FileFormModel;
 import org.usth.ict.ulake.common.model.folder.FileModel;
-import org.usth.ict.ulake.common.service.DashboardService;
+import org.usth.ict.ulake.common.service.CoreService;
+import org.usth.ict.ulake.textr.clients.DashboardRestClient;
 import org.usth.ict.ulake.textr.clients.FileRestClient;
 import org.usth.ict.ulake.textr.models.IndexFiles;
 import org.usth.ict.ulake.textr.models.IndexingStatus;
@@ -23,11 +25,15 @@ import org.usth.ict.ulake.textr.services.engines.IndexSearchEngineV2;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.transaction.Transactional;
+import javax.ws.rs.core.StreamingOutput;
+import java.io.InputStream;
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 
 @ApplicationScoped
-@Transactional(Transactional.TxType.REQUIRED)
+@Transactional
 public class IndexFilesService {
     
     Logger logger = LoggerFactory.getLogger(IndexFilesService.class);
@@ -40,11 +46,15 @@ public class IndexFilesService {
     
     @Inject
     @RestClient
-    DashboardService dashboardService;
+    DashboardRestClient dashboardService;
     
     @Inject
     @RestClient
     FileRestClient fileService;
+    
+    @Inject
+    @RestClient
+    CoreService coreService;
     
     @Inject
     ObjectMapper mapper;
@@ -52,21 +62,24 @@ public class IndexFilesService {
     @Inject
     JsonWebToken jwt;
     
-    public ServiceResponseBuilder<?> upload(String bearer, FileFormModel form) {
-        LakeHttpResponse<FileModel> response = dashboardService.newFile(bearer, form);
+    public Uni<ServiceResponseBuilder<?>> upload(String bearer, FileFormModel form) {
+        Uni<LakeHttpResponse<FileModel>> response = dashboardService.newFile(bearer, form);
         
-        if (response.getCode() != 200)
-            return new ServiceResponseBuilder<>(response.getCode(), response.getMsg(), response.getResp());
-        
-        FileModel fileModel = response.getResp();
-        String coreId = fileModel.cid;
-        Long fileId = fileModel.id;
-        Long size = fileModel.size;
-        
-        IndexFiles indexFiles = new IndexFiles(coreId, fileId, size, IndexingStatus.STATUS_SCHEDULED);
-        indexFilesRepo.save(indexFiles);
-        
-        return new ServiceResponseBuilder<>(200, "File scheduled");
+        return response.onItem().transform(lakeResponse -> {
+            if (lakeResponse.getCode() != 200)
+                return new ServiceResponseBuilder<>(lakeResponse.getCode(), lakeResponse.getMsg(),
+                                                    lakeResponse.getResp());
+            
+            FileModel fileModel = lakeResponse.getResp();
+            String coreId = fileModel.cid;
+            Long fileId = fileModel.id;
+            Long size = fileModel.size;
+            
+            IndexFiles indexFiles = new IndexFiles(coreId, fileId, size, IndexingStatus.STATUS_SCHEDULED);
+            indexFilesRepo.save(indexFiles);
+            
+            return new ServiceResponseBuilder<>(200, "File scheduled");
+        }).onFailure().recoverWithItem(e -> new ServiceResponseBuilder<>(500, e.getMessage()));
     }
     
     public ServiceResponseBuilder<List<FileModel>> listAllByStatus(IndexingStatus status, int page, int size) {
@@ -137,65 +150,65 @@ public class IndexFilesService {
         return new ServiceResponseBuilder<>(200, documentResponses);
     }
     
-    public ServiceResponseBuilder<?> delete(String bearer, Long id) {
-        IndexFiles indexFile = indexFilesRepo.findById(id).orElse(null);
+    public ServiceResponseBuilder<?> delete(String bearer, Long fid) {
+        IndexFiles indexFile = indexFilesRepo.findByFileId(fid).orElse(null);
         
         if (indexFile == null)
             return new ServiceResponseBuilder<>(404, "File not found");
         
-        if (this.nonUpdatable(indexFile.getStatus(), IndexingStatus.STATUS_IGNORED))
+        if (indexFile.getStatus() == IndexingStatus.STATUS_IGNORED)
             return new ServiceResponseBuilder<>(400, "File already deleted");
         
-        Long fid = indexFile.getFileId();
         LakeHttpResponse<FileModel> response = fileService.deleteFile(bearer, fid);
         
         if (response.getCode() != 200)
             return new ServiceResponseBuilder<>(response.getCode(), response.getMsg(), response.getResp());
         
-        indexFilesRepo.updateStatusById(id, IndexingStatus.STATUS_IGNORED);
+        indexFilesRepo.updateStatusByFileId(fid, IndexingStatus.STATUS_IGNORED);
         return new ServiceResponseBuilder<>(200, "File deleted");
     }
     
-    private boolean nonUpdatable(IndexingStatus status, IndexingStatus targetStatus) {
-        if (targetStatus == IndexingStatus.STATUS_INDEXED)
-            return true;
+    public ServiceResponseBuilder<?> reindex(Long fid) {
+        IndexFiles indexFile = indexFilesRepo.findByFileId(fid).orElse(null);
         
-        // Update from ignored to scheduled
-        if (targetStatus == IndexingStatus.STATUS_SCHEDULED && status != IndexingStatus.STATUS_IGNORED)
-            return true;
+        if (indexFile == null || indexFile.getStatus() == IndexingStatus.STATUS_IGNORED)
+            return new ServiceResponseBuilder<>(400, "File not found");
         
-        // Update from scheduled and indexed to ignored
-        return targetStatus == IndexingStatus.STATUS_IGNORED && status == IndexingStatus.STATUS_IGNORED;
-    }
-    
-    public ServiceResponseBuilder<?> restore(Long id) {
-        IndexFiles indexFile = indexFilesRepo.findById(id).orElse(null);
-        
-        if (indexFile == null)
-            return new ServiceResponseBuilder<>(404, "File not found");
-        
-        if (this.nonUpdatable(indexFile.getStatus(), IndexingStatus.STATUS_SCHEDULED))
-            return new ServiceResponseBuilder<>(400, "File already restored");
-        
-        indexFilesRepo.updateStatusById(id, IndexingStatus.STATUS_SCHEDULED);
-        return new ServiceResponseBuilder<>(200, "File restored");
-    }
-    
-    public ServiceResponseBuilder<?> reindex(Long id) {
-        IndexFiles indexFile = indexFilesRepo.findByIdAndStatus(id, IndexingStatus.STATUS_INDEXED).orElse(null);
-        
-        if (indexFile == null)
-            return new ServiceResponseBuilder<>(400, "File not found or scheduled");
+        if (indexFile.getStatus() == IndexingStatus.STATUS_SCHEDULED)
+            return new ServiceResponseBuilder<>(400, "File already scheduled");
         
         String cid = indexFile.getCoreId();
         
         try {
             indexSearchEngine.deleteDoc(cid);
-            indexFilesRepo.updateStatusById(id, IndexingStatus.STATUS_SCHEDULED);
+            indexFilesRepo.updateStatusByFileId(fid, IndexingStatus.STATUS_SCHEDULED);
         } catch (Exception e) {
             logger.error("Reindex doc failed: ", e);
             return new ServiceResponseBuilder<>(500, "File reindex failed" + e.getMessage());
         }
         return new ServiceResponseBuilder<>(200, "File re-indexing in progress");
+    }
+    
+    public ServiceResponseBuilder<StreamingOutput> download(Long fid) {
+        IndexFiles file = indexFilesRepo.findByFileId(fid).orElse(null);
+        
+        if (file == null)
+            return new ServiceResponseBuilder<>(404, "File not found");
+        
+        String bearer = "bearer " + jwt.getRawToken();
+        
+        FileModel fileModel = this.getFilesInfo(bearer, List.of(new IndexFiles[]{file})).getResp().get(0);
+        String encodedFilename = URLEncoder.encode(fileModel.name, StandardCharsets.UTF_8).replace("+", "%20");
+        
+        StreamingOutput streamingOutput;
+        try {
+            InputStream stream = coreService.objectDataByFileId(fid, bearer);
+            streamingOutput = stream::transferTo;
+        } catch (Exception e) {
+            logger.error("Download doc failed: ", e);
+            return new ServiceResponseBuilder<>(500, "Download failed: " + e.getMessage());
+        }
+        
+        return new ServiceResponseBuilder<>(200, encodedFilename, streamingOutput);
     }
 }
